+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.7-test.16",
+  "version": "1.7.7-test.18",
   "publishConfig": {
     "access": "public"
   },
