+++ lib/src/components/carousel/index.ts
