+++ lib/src/components/carousel/useScreenSize.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
 // Define Tailwind CSS breakpoints
 const breakpoints = {
