+++ lib/src/components/carousel/variants.ts
